Moved video capture handling to utilvideo.h/cpp
