+++ Client/qtTeamTalk/mainwindow.cpp
@@ -47,6 +47,7 @@
 #include "userdesktopdlg.h"
 #include "appinfo.h"
 #include "bearwarelogindlg.h"
+#include "utilvideo.h"
 
 #include <QMessageBox>
 #include <QInputDialog>
